+++ src/service/src/models/mod.rs
@@ -1,6 +1,6 @@
 // prose-pod-api
 //
-// Copyright: 2024, Rémi Bardon <remi@remibardon.name>
+// Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 mod data;
@@ -8,12 +8,14 @@ pub mod durations;
 mod email_address;
 pub mod sea_orm;
 mod serializable_secret_string;
+mod url;
 pub mod xmpp;
 
 pub use data::*;
 pub use durations::*;
 pub use email_address::*;
 pub use serializable_secret_string::*;
+pub use url::*;
 pub use xmpp::{
     jid, BareJid, FullJid, JidDomain, JidNode, XmppConnectionType, XmppDirectionalConnectionType,
     JID,